pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0", "wheel"]

[project]
name = "xclipboard"
version = "1.0.0"
description = "Cross-platform clipboard manipulation module for Python"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "XClipboard Contributors"}
]
keywords = ["clipboard", "cross-platform", "windows", "macos", "linux", "copy", "paste"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
image = [
    "Pillow>=9.0",
]
macos = [
    "pyobjc-framework-Cocoa>=9.0",
]

[project.urls]
Homepage = "https://github.com/example/xclipboard"
Repository = "https://github.com/example/xclipboard"
Documentation = "https://github.com/example/xclipboard#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["xclipboard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["xclipboard"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
